pyproject.toml
[project]
name = "avatar-agent"
version = "0.1.0"
description = "AI 数字人口播视频自动化生成工具"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "yt-dlp>=2024.1.0",
    "playwright>=1.45.0",
    "Pillow>=10.0.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.5.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
